Add files via upload
・small fix
・implementation of projection for gradient and hessian

+++ biaspotpy/keep_angle_potential.py
@@ -40,19 +40,19 @@ class StructKeepAnglePotential:
                              self.config["keep_angle_v2_fragm3"],
                              
         """
-        fragm_1_center = torch.tensor([0.0, 0.0, 0.0], dtype=torch.float32, requires_grad=True)
+        fragm_1_center = torch.tensor([0.0, 0.0, 0.0], dtype=torch.float64, requires_grad=True)
         for i in self.config["keep_angle_v2_fragm1"]:
             fragm_1_center = fragm_1_center + geom_num_list[i-1]
         
         fragm_1_center = fragm_1_center / len(self.config["keep_angle_v2_fragm1"])
         
-        fragm_2_center = torch.tensor([0.0, 0.0, 0.0], dtype=torch.float32, requires_grad=True)
+        fragm_2_center = torch.tensor([0.0, 0.0, 0.0], dtype=torch.float64, requires_grad=True)
         for i in self.config["keep_angle_v2_fragm2"]:
             fragm_2_center = fragm_2_center + geom_num_list[i-1]
         
         fragm_2_center = fragm_2_center / len(self.config["keep_angle_v2_fragm2"]) 
             
-        fragm_3_center = torch.tensor([0.0, 0.0, 0.0], dtype=torch.float32, requires_grad=True)
+        fragm_3_center = torch.tensor([0.0, 0.0, 0.0], dtype=torch.float64, requires_grad=True)
         for i in self.config["keep_angle_v2_fragm3"]:
             fragm_3_center = fragm_3_center + geom_num_list[i-1]
