+++ biaspotpy/keep_outofplain_angle_potential.py
@@ -41,25 +41,25 @@ class StructKeepOutofPlainAnglePotential:
                              self.config["keep_out_of_plain_angle_v2_fragm4"],
                              
         """
-        fragm_1_center = torch.tensor([0.0, 0.0, 0.0], dtype=torch.float32, requires_grad=True)
+        fragm_1_center = torch.tensor([0.0, 0.0, 0.0], dtype=torch.float64, requires_grad=True)
         for i in self.config["keep_out_of_plain_angle_v2_fragm1"]:
             fragm_1_center = fragm_1_center + geom_num_list[i-1]
         
         fragm_1_center = fragm_1_center / len(self.config["keep_out_of_plain_angle_v2_fragm1"])
         
-        fragm_2_center = torch.tensor([0.0, 0.0, 0.0], dtype=torch.float32, requires_grad=True)
+        fragm_2_center = torch.tensor([0.0, 0.0, 0.0], dtype=torch.float64, requires_grad=True)
         for i in self.config["keep_out_of_plain_angle_v2_fragm2"]:
             fragm_2_center = fragm_2_center + geom_num_list[i-1]
         
         fragm_2_center = fragm_2_center / len(self.config["keep_out_of_plain_angle_v2_fragm2"]) 
             
-        fragm_3_center = torch.tensor([0.0, 0.0, 0.0], dtype=torch.float32, requires_grad=True)
+        fragm_3_center = torch.tensor([0.0, 0.0, 0.0], dtype=torch.float64, requires_grad=True)
         for i in self.config["keep_out_of_plain_angle_v2_fragm3"]:
             fragm_3_center = fragm_3_center + geom_num_list[i-1]
         
         fragm_3_center = fragm_3_center / len(self.config["keep_out_of_plain_angle_v2_fragm3"])   
 
-        fragm_4_center = torch.tensor([0.0, 0.0, 0.0], dtype=torch.float32, requires_grad=True)
+        fragm_4_center = torch.tensor([0.0, 0.0, 0.0], dtype=torch.float64, requires_grad=True)
         for i in self.config["keep_out_of_plain_angle_v2_fragm4"]:
             fragm_4_center = fragm_4_center + geom_num_list[i-1]
